requirements.txt
beautifulsoup4==4.12.3
langchain==0.1.16
langchain_community==0.0.34
langchain_core==0.1.45
langchain_openai==0.1.3
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.33.0
faiss-cpu==1.8.0
